Move gas estimation logic to eth_estimateGas
