+++ crates/module-system/module-implementations/sov-evm/tests/integration/contracts.rs
@@ -52,7 +52,7 @@ fn test_invalid_contract_execution() {
         let cfg_env =
             CfgEnv::new_with_spec(SpecId::SHANGHAI).with_chain_id(config_value!("CHAIN_ID"));
         let tx = convert_to_transaction_signed(signed_eth_tx).unwrap();
-        let tx_env = create_tx_env(&tx, account.address(), 1, 1_000_000);
+        let tx_env = create_tx_env(&tx, account.address(), 1);
         let result =
             executor::transact_commit(evm_db, &BlockEnv::default(), tx_env, cfg_env).unwrap();
         assert!(matches!(result, ExecutionResult::Revert { .. }));
